:truck: Refactor the Field Configuration Service
1. Moved the Field Configuration Item method to a separated service under the Field Configuration service.

2. Replicated the v2 changes on the v3 module.
