+++ jira/v2/jira.go
@@ -88,8 +88,11 @@ func New(httpClient *http.Client, site string) (client *Client, err error) {
 			client: client,
 		},
 		Field: &FieldService{
-			client:        client,
-			Configuration: &FieldConfigurationService{client: client},
+			client: client,
+			Configuration: &FieldConfigurationService{
+				client: client,
+				Item:   &FieldConfigurationItemService{client: client},
+			},
 
 			Context: &FieldContextService{
 				client: client,
