+++ jira/v3/issueFieldConfiguration.go
@@ -9,11 +9,14 @@ import (
 	"strconv"
 )
 
-type FieldConfigurationService struct{ client *Client }
+type FieldConfigurationService struct {
+	client *Client
+	Item   *FieldConfigurationItemService
+}
 
 // Gets Returns a paginated list of all field configurations.
 // Docs: https://docs.go-atlassian.io/jira-software-cloud/issues/fields/configuration#get-all-field-configurations
-// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-field-configurations/#api-rest-api-3-fieldconfiguration-get
+// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v2/api-group-issue-field-configurations/#api-rest-api-2-fieldconfiguration-get
 func (f *FieldConfigurationService) Gets(ctx context.Context, ids []int, isDefault bool, startAt, maxResults int) (
 	result *models.FieldConfigurationPageScheme, response *ResponseScheme, err error) {
 
@@ -45,24 +48,34 @@ func (f *FieldConfigurationService) Gets(ctx context.Context, ids []int, isDefau
 	return
 }
 
-// Items Returns a paginated list of all fields for a configuration.
-// Docs: https://docs.go-atlassian.io/jira-software-cloud/issues/fields/configuration#get-field-configuration-items
-// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-field-configurations/#api-rest-api-3-fieldconfiguration-id-fields-get
-func (f *FieldConfigurationService) Items(ctx context.Context, fieldConfigurationID, startAt, maxResults int) (
-	result *models.FieldConfigurationItemPageScheme, response *ResponseScheme, err error) {
+// Create creates a field configuration. The field configuration is created with the same field properties as the
+// default configuration, with all the fields being optional.
+// This operation can only create configurations for use in company-managed (classic) projects.
+// EXPERIMENTAL
+func (f *FieldConfigurationService) Create(ctx context.Context, name, description string) (result *models.FieldConfigurationScheme,
+	response *ResponseScheme, err error) {
 
-	params := url.Values{}
-	params.Add("startAt", strconv.Itoa(startAt))
-	params.Add("maxResults", strconv.Itoa(maxResults))
+	if name == "" {
+		return nil, nil, models.ErrNoFieldConfigurationNameError
+	}
 
-	var endpoint = fmt.Sprintf("rest/api/3/fieldconfiguration/%v/fields?%v", fieldConfigurationID, params.Encode())
+	payload := struct {
+		Name        string `json:"name"`
+		Description string `json:"description,omitempty"`
+	}{
+		Name:        name,
+		Description: description,
+	}
+	payloadAsReader, _ := transformStructToReader(&payload)
+	endpoint := "rest/api/3/fieldconfiguration"
 
-	request, err := f.client.newRequest(ctx, http.MethodGet, endpoint, nil)
+	request, err := f.client.newRequest(ctx, http.MethodPost, endpoint, payloadAsReader)
 	if err != nil {
 		return
 	}
 
 	request.Header.Set("Accept", "application/json")
+	request.Header.Set("Content-Type", "application/json")
 
 	response, err = f.client.call(request, &result)
 	if err != nil {
@@ -72,9 +85,76 @@ func (f *FieldConfigurationService) Items(ctx context.Context, fieldConfiguratio
 	return
 }
 
+// Update updates a field configuration. The name and the description provided in the request override the existing values.
+// This operation can only update configurations used in company-managed (classic) projects.
+// EXPERIMENTAL
+func (f *FieldConfigurationService) Update(ctx context.Context, fieldConfigurationID int, name, description string) (
+	response *ResponseScheme, err error) {
+
+	if fieldConfigurationID == 0 {
+		return nil, models.ErrNoFieldConfigurationIDError
+	}
+
+	if name == "" {
+		return nil, models.ErrNoFieldConfigurationNameError
+	}
+
+	payload := struct {
+		Name        string `json:"name"`
+		Description string `json:"description,omitempty"`
+	}{
+		Name:        name,
+		Description: description,
+	}
+
+	payloadAsReader, _ := transformStructToReader(&payload)
+	endpoint := fmt.Sprintf("rest/api/3/fieldconfiguration/%v", fieldConfigurationID)
+
+	request, err := f.client.newRequest(ctx, http.MethodPut, endpoint, payloadAsReader)
+	if err != nil {
+		return
+	}
+
+	request.Header.Set("Accept", "application/json")
+	request.Header.Set("Content-Type", "application/json")
+
+	response, err = f.client.call(request, nil)
+	if err != nil {
+		return
+	}
+
+	return
+}
+
+// Delete deletes a field configuration.
+// This operation can only delete configurations used in company-managed (classic) projects.
+// EXPERIMENTAL
+func (f *FieldConfigurationService) Delete(ctx context.Context, fieldConfigurationID int) (response *ResponseScheme, err error) {
+
+	if fieldConfigurationID == 0 {
+		return nil, models.ErrNoFieldConfigurationIDError
+	}
+
+	endpoint := fmt.Sprintf("rest/api/3/fieldconfiguration/%v", fieldConfigurationID)
+
+	request, err := f.client.newRequest(ctx, http.MethodDelete, endpoint, nil)
+	if err != nil {
+		return
+	}
+
+	request.Header.Set("Accept", "application/json")
+
+	response, err = f.client.call(request, nil)
+	if err != nil {
+		return
+	}
+
+	return
+}
+
 // Schemes Returns a paginated list of field configuration schemes.
 // Docs: https://docs.go-atlassian.io/jira-software-cloud/issues/fields/configuration#get-all-field-configuration-schemes
-// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-field-configurations/#api-rest-api-3-fieldconfigurationscheme-get
+// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v2/api-group-issue-field-configurations/#api-rest-api-2-fieldconfigurationscheme-get
 func (f *FieldConfigurationService) Schemes(ctx context.Context, IDs []int, startAt, maxResults int) (
 	result *models.FieldConfigurationSchemePageScheme, response *ResponseScheme, err error) {
 
@@ -105,7 +185,7 @@ func (f *FieldConfigurationService) Schemes(ctx context.Context, IDs []int, star
 
 // IssueTypeItems Returns a paginated list of field configuration issue type items.
 // Docs: https://docs.go-atlassian.io/jira-software-cloud/issues/fields/configuration#get-field-configuration-issue-type-items
-// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-field-configurations/#api-rest-api-3-fieldconfigurationscheme-mapping-get
+// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v2/api-group-issue-field-configurations/#api-rest-api-2-fieldconfigurationscheme-mapping-get
 func (f *FieldConfigurationService) IssueTypeItems(ctx context.Context, fieldConfigIDs []int, startAt, maxResults int) (
 	result *models.FieldConfigurationIssueTypeItemPageScheme, response *ResponseScheme, err error) {
 
@@ -135,7 +215,7 @@ func (f *FieldConfigurationService) IssueTypeItems(ctx context.Context, fieldCon
 
 // SchemesByProject Returns a paginated list of field configuration schemes and, for each scheme, a list of the projects that use it.
 // Docs: https://docs.go-atlassian.io/jira-software-cloud/issues/fields/configuration#get-field-configuration-schemes-for-projects
-// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-field-configurations/#api-rest-api-3-fieldconfigurationscheme-project-get
+// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v2/api-group-issue-field-configurations/#api-rest-api-2-fieldconfigurationscheme-project-get
 func (f *FieldConfigurationService) SchemesByProject(ctx context.Context, projectIDs []int, startAt, maxResults int) (
 	result *models.FieldConfigurationSchemeProjectPageScheme, response *ResponseScheme, err error) {
 
@@ -166,7 +246,7 @@ func (f *FieldConfigurationService) SchemesByProject(ctx context.Context, projec
 
 // Assign assigns a field configuration scheme to a project.
 // If the field configuration scheme ID is null, the operation assigns the default field configuration scheme.
-// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-field-configurations/#api-rest-api-3-fieldconfigurationscheme-project-put
+// Official Docs: https://developer.atlassian.com/cloud/jira/platform/rest/v2/api-group-issue-field-configurations/#api-rest-api-2-fieldconfigurationscheme-project-put
 func (f *FieldConfigurationService) Assign(ctx context.Context, fieldConfigurationSchemeID, projectID string) (response *ResponseScheme, err error) {
 
 	if len(projectID) == 0 {
